+++ spec/rails_app/app/components/grid/scoped.rb
@@ -3,6 +3,6 @@ class Grid::Scoped < Netzke::Basepack::Grid
     c.model = "Book"
     super
     c.scope = lambda {|r| r.where(author_id: Author.first.try(:id))}
-    c.strong_default_attrs = {:author_id => Author.first.id}
+    c.strong_default_attrs = {:author_id => Author.first.id, notes: 'Fixed'}
   end
 end