+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 0.12.7 - WIP
+# Fix the strong_default_attrs option support
 
 # 0.12.6 - 2015-07-11
 * Regression: show cell editor on adding/editing a record in grid